requirements.txt
Flask==2.3.3
openai-whisper==20231117
numpy==1.24.3